开始 pdf 划选高亮显示

+++ pdflibrary/src/main/java/com/example/pdflibrary/util/ResultCoordinate.java
@@ -1,6 +1,6 @@
 package com.example.pdflibrary.util;
 
 public class ResultCoordinate {
-    public Double x;
-    public Double y;
+    public Float x;
+    public Float y;
 }
